- correct bug on Real eq constraint manager - create DeterminsticIndicedList object to record objects in Model - modify CPModel to declare DeterministicIndicedList - delete list of indexes in Constraint and Variable (and sub classes) - simplify IIndex interface - separate build time and preprocess time in XmlModel - add new constructor of CPModel to include intiale size of data structure - modify XmlModel to include new constructor of CPModel
git-svn-id: svn://svn.code.sf.net/p/choco/code/trunk@1096 5779b9a9-2447-0410-887e-bb98d57c6de4

+++ choco-kernel/src/main/java/choco/kernel/common/IndexFactory.java
@@ -33,18 +33,18 @@ package choco.kernel.common;
 */
 public class IndexFactory {
 
-    public int index = 33;
+    public long index = 33;
 
     /**
      * Return a unique index
      * @return
      */
-    public final int getIndex(){
+    public final long getIndex(){
         return ++index;
     }
 
 
-    public static int base = (int)System.currentTimeMillis();
+    public static long base = System.currentTimeMillis();
 
     /**
      * STATIC = STRONGLY JVM DEPENDANT!!
@@ -54,7 +54,7 @@ public class IndexFactory {
      * Return a unique id
      * @return
      */
-    public static synchronized int getId(){
+    public static synchronized long getId(){
         return ++base;
     }
